requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15,<4.0
django-cors-headers>=4.5,<5.0
whitenoise>=6.6,<7.0
gunicorn>=21,<26

django-oauth-toolkit>=3.0,<4.0
drf-social-oauth2>=3.0,<4.0
social-auth-app-django>=5.0,<6.0
social-auth-core>=4.5,<5.0
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0
oauthlib>=3.2,<4.0
requests-oauthlib>=2.0,<3.0
python3-openid>=3.2,<4.0

psycopg2-binary>=2.9,<3.0
pgvector==0.4.2

google-api-core[grpc]>=2.25,<3.0
google-api-python-client>=2.100,<3.0
google-auth>=2.25,<3.0
google-auth-httplib2>=0.2,<1.0
googleapis-common-protos>=1.60,<2.0
google-ai-generativelanguage>=0.6,<1.0
google-genai>=1.60,<2.0

grpcio>=1.60,<2.0

httpx>=0.27,<1.0
aiohttp>=3.9,<4.0
anyio>=4.0,<5.0
websockets>=12,<17

python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
urllib3>=2.0,<3.0
certifi>=2024.0
charset-normalizer>=3.0,<4.0
idna>=3.4,<4.0

cryptography>=42,<47
cffi>=1.16,<3.0

numpy>=1.26,<3.0

tqdm>=4.60,<5.0
tenacity>=8.2,<10.0
packaging>=23.0
typing_extensions>=4.8,<5.0
pydantic>=2.5,<3.0

livekit-api>=1.0,<2.0
livekit-protocol>=1.0,<2.0

cloudinary>=1.40,<2.0